Develop

+++ alignn/ff/ff.py
@@ -25,11 +25,7 @@ from ase import units
 from ase.md import Langevin
 from ase.md.npt import NPT
 from ase.md.andersen import Andersen
-import ase.calculators.calculator
-from ase.stress import full_3x3_to_voigt_6_stress
 from jarvis.db.jsonutils import loadjson
-from alignn.graphs import Graph
-from alignn.models.alignn_atomwise import ALIGNNAtomWise, ALIGNNAtomWiseConfig
 from jarvis.analysis.defects.vacancy import Vacancy
 import numpy as np
 from alignn.pretrained import get_prediction
@@ -49,7 +45,13 @@ from ase.cell import Cell
 from matplotlib.gridspec import GridSpec
 from sklearn.metrics import mean_absolute_error
 from tqdm import tqdm
-import torch
+
+# import torch
+from alignn.ff.calculators import (
+    AlignnAtomwiseCalculator,
+)  # Import from new location
+
+__all__ = ["AlignnAtomwiseCalculator"]
 
 try:
     from gpaw import GPAW, PW
@@ -61,7 +63,16 @@ except Exception:
 
 
 def get_all_models():
-    json_path = os.path.join(os.path.dirname(__file__), "all_models_ff.json")
+    json_path = os.path.join(
+        os.path.dirname(__file__), "all_models_alignn_atomwise.json"
+    )
+    return loadjson(json_path)
+
+
+def get_all_models_prop():
+    json_path = os.path.join(
+        os.path.dirname(__file__), "all_models_alignn.json"
+    )
     return loadjson(json_path)
 
 
@@ -111,6 +122,71 @@ def get_figshare_model_ff(
         # print("Loading the zipfile...", zipfile.ZipFile(path).namelist())
         data = zipfile.ZipFile(path).read(tmp)
 
+        # new_file, filename = tempfile.mkstemp()
+        filename = os.path.join(dir_path, filename)
+        with open(filename, "wb") as f:
+            f.write(data)
+        filename = os.path.join(dir_path, "config.json")
+        with open(filename, "wb") as f:
+            f.write(config)
+        os.remove(path)
+    # print("Using model file", url, "from ", chks)
+    # print("Path", os.path.abspath(path))
+    # print("Config", os.path.abspath(cfg))
+    return dir_path
+
+
+def get_figshare_model_prop(
+    model_name="jv_mbj_bandgap_alignn", dir_path=None, filename="best_model.pt"
+):
+    """Get ALIGNN-FF torch models from figshare."""
+    all_models_prop = get_all_models_prop()
+    # https://doi.org/10.6084/m9.figshare.23695695
+    if dir_path is None:
+        dir_path = str(os.path.join(os.path.dirname(__file__), model_name))
+    # cwd=os.getcwd()
+    dir_path = os.path.abspath(dir_path)
+    if not os.path.exists(dir_path):
+        os.makedirs(dir_path)
+    # os.chdir(dir_path)
+    url = all_models_prop[model_name]
+    zfile = model_name + ".zip"
+    path = str(os.path.join(dir_path, zfile))
+    # path = str(os.path.join(os.path.dirname(__file__), zfile))
+    print("dir_path", dir_path)
+    best_path = os.path.join(dir_path, filename)
+    if not os.path.exists(best_path):
+        response = requests.get(url, stream=True)
+        total_size_in_bytes = int(response.headers.get("content-length", 0))
+        block_size = 1024  # 1 Kibibyte
+        progress_bar = tqdm(
+            total=total_size_in_bytes, unit="iB", unit_scale=True
+        )
+        with open(path, "wb") as file:
+            for data in response.iter_content(block_size):
+                progress_bar.update(len(data))
+                file.write(data)
+        progress_bar.close()
+        zp = zipfile.ZipFile(path)
+        names = zp.namelist()
+        chks = []
+        cfg = []
+        for i in names:
+            if "checkpoint_" in i and "pt" in i:
+                tmp = i
+                # fname = i
+                chks.append(i)
+            if filename in i:
+                tmp = i
+                chks.append(i)
+
+            if "config.json" in i:
+                cfg = i
+
+        config = zipfile.ZipFile(path).read(cfg)
+        # print("Loading the zipfile...", zipfile.ZipFile(path).namelist())
+        data = zipfile.ZipFile(path).read(tmp)
+        # print('dir_path',dir_path,filename)
         # new_file, filename = tempfile.mkstemp()
         filename = os.path.join(dir_path, filename)
         with open(filename, "wb") as f:
@@ -190,140 +266,6 @@ def ase_to_atoms(ase_atoms):
         #         pbc=True,
         cartesian=True,
     )
-
-
-ignore_bad_restart_file = ase.calculators.calculator.Calculator._deprecated
-
-
-class AlignnAtomwiseCalculator(ase.calculators.calculator.Calculator):
-    """Module for ASE Calculator interface."""
-
-    def __init__(
-        self,
-        restart=None,
-        ignore_bad_restart_file=ignore_bad_restart_file,
-        label=None,
-        include_stress=True,
-        intensive=True,
-        atoms=None,
-        directory=".",
-        device=None,
-        model=None,
-        config=None,
-        path=None,
-        model_filename="best_model.pt",
-        config_filename="config.json",
-        output_dir=None,
-        batch_stress=True,
-        force_mult_natoms=False,
-        force_mult_batchsize=True,
-        force_multiplier=1,
-        stress_wt=0.05,
-    ):
-        """Initialize class."""
-        super(AlignnAtomwiseCalculator, self).__init__(
-            restart, ignore_bad_restart_file, label, atoms, directory
-        )  # , **kwargs
-        self.model = model
-        self.device = device
-        self.intensive = intensive
-        self.config = config
-        self.include_stress = include_stress
-        self.stress_wt = stress_wt
-        self.force_mult_natoms = force_mult_natoms
-        self.force_mult_batchsize = force_mult_batchsize
-        self.force_multiplier = force_multiplier
-        if path is None and model is None:
-            path = default_path()
-        if self.config is None:
-            config = loadjson(os.path.join(path, config_filename))
-            self.config = config
-        if self.force_mult_natoms:
-            self.config["model"]["force_mult_natoms"] = True
-        if self.include_stress:
-            self.implemented_properties = ["energy", "forces", "stress"]
-            if self.config["model"]["stresswise_weight"] == 0:
-                self.config["model"]["stresswise_weight"] = 0.1
-        else:
-            self.implemented_properties = ["energy", "forces"]
-
-        if batch_stress is not None:
-            self.config["model"]["batch_stress"] = batch_stress
-        import torch
-
-        if self.device is None:
-            self.device = torch.device(
-                "cuda" if torch.cuda.is_available() else "cpu"
-            )
-        if self.model is None:
-
-            if self.config["model"]["name"] == "alignn_atomwise":
-                model = ALIGNNAtomWise(
-                    ALIGNNAtomWiseConfig(**self.config["model"])
-                )
-            model.state_dict()
-            model.load_state_dict(
-                torch.load(
-                    os.path.join(path, model_filename),
-                    map_location=self.device,
-                )
-            )
-            model.to(device)
-            model.eval()
-            self.model = model
-        else:
-            model = self.model
-
-    def calculate(self, atoms, properties=None, system_changes=None):
-        """Calculate properties."""
-        j_atoms = ase_to_atoms(atoms)
-        num_atoms = j_atoms.num_atoms
-        g, lg = Graph.atom_dgl_multigraph(
-            j_atoms,
-            neighbor_strategy=self.config["neighbor_strategy"],
-            cutoff=self.config["cutoff"],
-            max_neighbors=self.config["max_neighbors"],
-            atom_features=self.config["atom_features"],
-            use_canonize=self.config["use_canonize"],
-        )
-
-        if self.config["model"]["alignn_layers"] > 0:
-            result = self.model(
-                (
-                    g.to(self.device),
-                    lg.to(self.device),
-                    torch.tensor(atoms.cell)
-                    .type(torch.get_default_dtype())
-                    .to(self.device),
-                )
-            )
-        else:
-            result = self.model(
-                (g.to(self.device, torch.tensor(atoms.cell).to(self.device)))
-            )
-        forces = forces = (
-            result["grad"].detach().cpu().numpy() * self.force_multiplier
-        )
-        stress = (
-            full_3x3_to_voigt_6_stress(
-                result["stresses"][:3].reshape(3, 3).detach().cpu().numpy()
-            )
-            * self.stress_wt
-            / 160.21766208
-        )
-        energy = result["out"].detach().cpu().numpy()
-        if self.intensive:
-            energy *= num_atoms
-        if self.force_mult_natoms:
-            forces *= num_atoms
-        if self.force_mult_batchsize:
-            forces *= self.config["batch_size"]
-
-        self.results = {
-            "energy": energy,
-            "forces": forces,
-            "stress": stress,
-        }
 
 
 class ForceField(object):
